requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-texttospeech==2.23.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
idna==3.10
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyPDF2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.3.0
